[ui][fix] Dynamic wait for mysql database start, cleaning openshift bdd steps in utils
