+++ ui-tests/src/test/resources/features/integrations/db-to-db-mysql.feature
@@ -8,7 +8,7 @@ Feature: Integration - DB to DB mysql
     Given log into the Syndesis
     Given clean MySQL server
     Given deploy MySQL server
-    And   sleep for "120000" ms
+    And wait until mysql database starts
     Given create standard table schema on "mysql" driver
 
     When inserts into "contact" table on "mysql"
